Various fixes and optimizations to TENxMatrix objects
The "extract_array" method for TENxMatrixSeed objects had several bugs
leading to the wrong data ending up in memory. Its implementation has been
revisited, fixed, and tested. It has also been optimized for the common
use cases. Depending on the size of the data that needs to be loaded, the
speedup can vary between 2x and 10x.

Also the examples in ?TENxMatrix have been expanded significantly.