HDF5Array objects now find their dimnames in the HDF5 file
writeHDF5Array() and as(x, "HDF5Array") know how to write the dimnames to
the HDF5 file, and the HDF5Array() constructor knows how to find them.
See ?writeHDF5Array for more information.

So that's it! This long overdue feature is finally implemented. It relies
on the HDF5 High-level library which was recently added to the Rhdf5lib
package by Mike Smith (in BioC 3.10). Thanks Mike!