Add 'as.sparse' argument to HDFArray/HDF5ArraySeed constructors
This allows the user to control wether the HDF5 dataset should be considered
sparse (and treated as such) or not. More precisely, when 'as.sparse' is set
to TRUE, blocks will be loaded as sparse objects during block processing,
which should lead to less memory usage and hopefully overall better
performance.