Add H5SparseMatrix class and constructor function
H5SparseMatrix is a DelayedMatrix subclass for representing and
operating on an HDF5 sparse matrix stored in CSR/CSC/Yale format.
This new class relies on and leverages the work that was done already
on H5SparseMatrixSeed objects.

This commit also replaces argument 'name' with argument 'layer' in the
H5ADMatrix() constructor.